added IKeyspaceNotificationOptions
